Add ApplicantProjectViewHandler
add ApplicantProjectViewHandler, do not allow user to add msg to other ppl's enquiry
